fixed problem with items not getting added to the cart from the wishlist page and added automotive parts to the database

+++ wishlist.php
@@ -1,3 +1,49 @@
+<?php
+@include 'config.php';
+
+session_start();
+
+$user_id = $_SESSION['user_id'];
+
+if (!isset($user_id)) {
+   header('location: login.php');
+};
+
+
+if(isset($_POST['add_to_cart'])){
+    $car_id = $_POST['car_id'];
+    $car_name = $_POST['car_name'];
+    $car_price = $_POST['car_price'];
+    $car_image = $_POST['car_image'];
+    $car_quantity = $_POST['car_quantity'];
+
+    // Debugging statement
+    echo "car ID: $car_id, Name: $car_name, Price: $car_price, Image: $car_image, Quantity: $car_quantity";
+
+    $check_cart_numbers = mysqli_query($conn, "SELECT * FROM `cart` WHERE name = '$car_name' AND user_id = '$user_id'") or die('query failed');
+
+    if(mysqli_num_rows($check_cart_numbers) > 0){
+        $message[] = 'already added to cart';
+    } else {
+        // Debugging statement
+        echo "Inserting into cart table";
+
+        mysqli_query($conn, "INSERT INTO `cart`(user_id, pid, name, price, quantity, image) VALUES('$user_id', '$car_id', '$car_name', '$car_price', '$car_quantity', '$car_image')") or die(mysqli_error($conn));
+
+        $message[] = 'product added to cart';
+    }
+}
+if(isset($_GET['delete'])){
+    $delete_id = $_GET['delete'];
+    mysqli_query($conn, "DELETE FROM `wishlist` WHERE id = '$delete_id'") or die('query failed');
+    header('location:wishlist.php');
+}
+
+if(isset($_GET['delete_all'])){
+    mysqli_query($conn, "DELETE FROM `wishlist` WHERE user_id = '$user_id'") or die('query failed');
+    header('location:wishlist.php');
+}
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -31,18 +77,17 @@
             while ($fetch_wishlist = mysqli_fetch_assoc($select_wishlist)) {
                 ?>
                 <form action="" method="POST" class="box">
-                    <a href="wishlist.php?delete=<?php echo $fetch_wishlist['id']; ?>"
-                       class="fas fa-times"
-                       onclick="return confirm('Delete this from wishlist?');"></a>
+                    <a href="wishlist.php?delete=<?php echo $fetch_wishlist['id']; ?>" class="fas fa-times" onclick="return confirm('delete this from wishlist?');"></a>
                     <img src="cars/<?php echo $fetch_wishlist['image']; ?>" alt="" class="image">
                     <div class="name"><?php echo $fetch_wishlist['name']; ?></div>
                     <div class="price">$<?php echo $fetch_wishlist['price']; ?>/-</div>
-                    <input type="hidden" name="product_id" value="<?php echo $fetch_wishlist['pid']; ?>">
-                    <input type="hidden" name="product_name" value="<?php echo $fetch_wishlist['name']; ?>">
-                    <input type="hidden" name="product_price" value="<?php echo $fetch_wishlist['price']; ?>">
-                    <input type="hidden" name="product_image" value="<?php echo $fetch_wishlist['image']; ?>">
+                    <input type="number" name="car_quantity" value="1" min="0" class="qty">
+                    <input type="hidden" name="car_id" value="<?php echo $fetch_wishlist['pid']; ?>">
+                    <input type="hidden" name="car_name" value="<?php echo $fetch_wishlist['name']; ?>">
+                    <input type="hidden" name="car_price" value="<?php echo $fetch_wishlist['price']; ?>">
+                    <input type="hidden" name="car_image" value="<?php echo $fetch_wishlist['image']; ?>">
                     <input type="submit" value="add to cart" name="add_to_cart" class="btn">
-                </form>
+               </form>
                 <?php
                 $grand_total += $fetch_wishlist['price'];
             }
